Make Program class public and add Main method
The Program class was updated to have a public access modifier,
allowing it to be used outside its assembly. A Main method was
added as the entry point of the application. Additionally, a
Console.Write statement was introduced to prompt the user to
enter their name, enabling basic user interaction.